+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release Notes
 
 - use "Exile" rather than "Banishment" for exiles
+- added global animation toggle for better performance
+- added record vote history toggle to session menu, and clear vote history button
+- add support for custom Fabled characters
+
+---
 
 ### Version 2.12.0
 - tweak reference sheet to better fit screen in single column layout